Author: Rex Kerr
Modified some of the data structures in mcell_structs.h to reflect
discussions over the past few months.

Added react_trig.c, which is a set of routines that will be called to
determine whether a particle undergoes a reaction (every timestep for
unimolecular, every collision for bimolecular, and every time hitting a
wall, for intersect).  These all return a pointer to a rxn data structure
indicating what may happen, or NULL if nothing will happen.