Author: Rex Kerr
Added react_outc.c, which handles outcomes of reactions.  Right now it
doesn't handle changes in identity very gracefully, and it doesn't update
subvolume membership properly, but it won't change much.  (It also isn't
debugged yet.)  Basic idea is that you call outcome_reactiontype and it
handles everything for the reactants, then calls a common function to create
products of the requested types and orientations.

Added functions to grid_utils to convert between grid index and u,v (surface)
coodinates.

And, of course, we have our mandatory changes to mcell_structs to facilitate
the new things in react_outc.