Tom Bartol:
In MCell-v2 numerous simulation state variables were stored in global
variables.

In MCell-v3 we'll store these as members of struct volume.
This commit adds to struct volume the state variables for the command-line
options and some (but probably not all) of the state variables required
to enable our first test of MCell-v3 involving release, diffusion, and
interaction of two diffusing species in unbounded space -- all directed
by a parsed MDL file.