Author: Rex Kerr
Piles of large and small bugfixes.  Diffusion code is now in and appears to
be working for free diffusion.  (My first run was rather hilarous--I had
forgotten to turn off time-skipping, so all particles instantly jumped the
maximum number of timesteps when they realized there was nothing to react
with.)

Added lots of sources to the makefile, including the reaction ones (since my
diffusion loop calls those).

There are still major bugs in scheduling, mostly because I'm doing it in an
overly complicated way.  This should be fixed next.  (diffuse_3D in
diffuse.c)

Also need to convert times in release data structure into units of
timesteps.