Author: Rex Kerr
Reworked scheduling scheme.  The new one is much more streamlined and does
away with one of the three doubles.  Now there is only t, the time that the
molecule will next be active, and t2, which is the amount of time until the
next unimolecular reaction.  If t advances by t2, the reaction occurs and t2
is rescheduled.

Added a pointer in the abstract_molecule struct so that the molecules can
keep track of where they were allocated from.  This does add 4 bytes per
molecule, but also makes cleanup much easier.