Author: Rex Kerr
Implemented a Monte-Carlo estimation scheme to deal with particle targets
that intersect subvolumes and walls.  Testing shows less that 1.5% error in
reaction rate even with fairly perverse cases (and the estimate of correct
reaction rate could be slightly off, so the true error may be less).

This fixes the problem mentioned in my last update.

Note: the code is not highly optimized and could probably be sped up by a
factor of two or three.

Also fixed a bug in spherical release sites where they were releasing in an
octant rather than in the entire sphere, and in reaction scaling with time,
which is now probably broken for long timesteps (I'll fix that later),
but at least is correct for minimum-length timesteps.