Author: Rex Kerr
Added a new keyword to the MDL syntax:
  INTERACTION_RADIUS = (number)
to allow the user to specify the radius of disks for molecule-molecule
collisions.  If it is not set, it defaults to the same area as specified by
the effector grid.

The code has been updated to use this value where applicable (except I
just realized I messed up the defaulting, so that'll come in properly
on my next commit).