Author: Rex Kerr
Added new release site shapes:
  CUBIC_RELEASE_SITE
  RECTANGULAR_RELEASE_SITE
  ELLIPTIC_RELEASE_SITE
CUBIC and RECTANGULAR are synonyms, as are SPHERICAL and ELLITPIC.  The
cubic versions unsurprisingly release in a cube instead of a sphere.

The release site diameter now can vary in each dimension:
  SITE_DIAMETER = [1,5,10000]
would produce a very elongated release site.

Note that geometric rotations do not work on release site shapes; they
always stay aligned with Cartesian coordinates and are unstretched.  (This
is a bug carried over from MCell version 2.  We should fix it eventually.)