Author: Rex Kerr
Fixed a bug where I was unwittingly comparing a double to NULL (which
escaped gcc3.3's notice).

Added the function create_grid to grid_util.  It creates memory for
an effector grid and initializes all the values, given a wall and a
suggested subvolume (or NULL if you don't know what subvolume to use).

Some updates have been made to start getting surface reactions to work,
but they're not complete yet.  The parser has been updated, though
so that if you use
  DIFFUSION_CONSTANT_2D = (number)
you will get a surface molecule or grid molecule (grid if (number) is
zero).
  DIFFUSION_CONSTANT
and
  DIFFUSION_CONSTANT_3D
mean the same thing.