Author: Rex Kerr
Reactions that have a grid molecule as a product will now replace an
existing grid molecule, but only if it is a reactant (otherwise the reaction will
fail to proceed).  It should be made a syntax error to have multiple grid
molecules on the right of a reaction (the old one should be destroyed).

Diffusion code now performs reactions between diffusing molecules and grid
molecules.  Surface molecules still cannot be interacted with.  This code
has not yet been tested, and therefore is probably buggy.

Modified estimate_disk routine (used for Monte Carlo estimation of
accessible area of interaction disk) to use a new routine, touch_wall.
This should be substantially faster.  It would probably be faster yet
if I didn't call estimate_disk at all for molecules not near anything
interesting.

Fixed a bug in the create_grid initialization routine.  (Tom: grab the new
one before trying to get it to work!)  Also fixed a (bitmask != NULL)
comparison (was supposed to be != 0).