Author: Rex Kerr
Added/modifed code for reactions between diffusing particles and grid
molecules so that it now seems to work properly in at least a gross way
(save for setting the orientation appropriately--I'm not sure I have the
conventions right, but that's trivial to modify).

The test.mdl file has a five-species system that shows reactions between
diffusing molecules and surfaces and between diffusing molecules and grid
molecules.