Tom Bartol:
  Adding molecules (aka effectors; aka grid molecules) to surface regions
is now working!!!  In the process of coding this we have revamped
and significantly streamlined the syntax.  Please note that the new syntax
has resulted in slight changes to the syntax for definition of surface classes
and surface regions.  Also note that right now molecules may only be added by
density -- adding molecules by number will be implemented shortly.
Here is how it works now:


It is now possible to define surface classes with molecule densities like
this:

DEFINE_SURFACE_CLASSES {
  surf_a {
    TRANSPARENT = ACh
    MOLECULE_DENSITY {
      ACf' = 500
      AChR.R, = 1500
    }
  }
}



It is possible to define surface regions with molecule densities
and set the surface class of a region from within a BOX or POLYGON_LIST
object definition like this:

box_1 BOX {
  CORNERS = [-0.1, -0.1, -0.1] , [0.1,0.1,0.1]
  DEFINE_SURFACE_REGIONS {
    reg1 {
      ELEMENT_LIST = [TOP,BOTTOM]
      SURFACE_CLASS = surf_a
      MOLECULE_DENSITY {
        AChR.R' = 1000
        AChR.AR' = 300
      }
    }
  }
}



It is also possible to define a surface region on a previously defined
BOX or POLYGON_LIST object like this:

DEFINE_SURFACE_REGIONS {
  box_1[reg2] {
    ELEMENT_LIST = [LEFT,RIGHT]
    SURFACE_CLASS = surf_a
    MOLECULE_DENSITY {
      ACf' = 250
    }
  }
}



And it is possible to modify the properties (except for the element_list)
of a previously defined region like this:

MODIFY_SURFACE_REGIONS {
  box_1[reg1] {
    SURFACE_CLASS = surf_b
    MOLECULE_DENSITY {
      ACf' = 2000
    }
  }
}