Author: Rex Kerr
Fixed a bug in grid2xyz.  Reaction geometries are broken; I'll fix them
soon.  In particular, surface/grid molecules don't release molecules
properly, so reactions of the form A + B <--> AB result in leaks.

Also, we need to multiply by the factor of two for one-sided reactions, as
we do in MCell2 (I presume).