Author: Rex Kerr
I ended up having to rewrite a very large portion of the reaction
preprocessor and of the reaction outcome code.  It seems to work again about
as well as it did before, and now it is actually possible to write reactions
that look like
  A' + B -> A, + B,
that work correctly.  Flags weren't doing the trick, so struct rxn no longer
has any flags.  The new method is a bit slower during initialization but
should be faster during simulation time.

I've also stopped up the primary leak of molecules that was occurring when a
grid_molecule fell apart into a grid_molecule and a ligand.

There are two new .mdl test files: mcell2_box_rx.mdl and mcell3_box_rx.mdl.
These are intended to be identical simulations in both versions.  The
answers currently come out rather different for me, but at least they both
run.