Author: Rex Kerr
MCell3 and MCell2 reaction rates for ligand-effector reactions are now the
same to within at least a few percent.  Bugs included a 4-argument location
in the MCell2 file, a copied reaction line that had zeros for orientation
instead of the real geometry, and another bug in interprogrammer
communications where Tom and I made different assumptions about the value of
orientations of grid molecules (I assume the only valid values are +-1).