Tom Bartol:
VIZ_DATA_OUTPUT for MODE = DX is now operational and ready for your
visualization enjoyment.  Syntax is the same as in MCell version 2.73.

I also added a new model to src/validate:  unbounded_mmi.mdl which is
a very simple test of molecule-molecule interactions in unbounded space.
The reaction simulated is A + B -> B.A  where B.A is frozen in space where
it forms.  Run it, try it, viz it!!!