Author: Rex Kerr
I forgot to mention a couple things in the last message.  User-defined
partitions work using the old syntax--these set the coarse partitions.  If
no partitions are given, the default ones are used.  If only some partition
axes are given, I'm not sure what happens.  I think you have to set all
three by hand or it ignores them.

Also, added the -fno-strict-aliasing option to makedefs.  This is more a
reminder to worry about C99 strict aliasing (and the way it's been
implemented in gcc) than an expectation that aliasing will cause problems
for us.  (Rather, it will probably cause warnings for us.)