Author: Rex Kerr
Various bug fixes.  As far as I can tell, MCell3 now replicates MCell2
results on simple systems.  CUSTOM_RK is a new viz mode that outputs one
four-column file.  The first column is the molecule ID (set in the states
block), and the second through fourth columns are the position.