Tom Bartol:
  REACTION_DATA_OUTPUT is now working for counting molecules in the world
including mathematical expressions involving counters.  The syntax is the
same as in MCell2 except that all count statements are assumed to be
"FOR_EACH_TIME_STEP" (i.e. they are a time-series of the counted quantitity)
and so this option is no longer needed.  If you want to
"CUMULATE_FOR_EACH_IME_STEPS" you can post-process your data easily in
xmgrace, matlab, mathematica, whatever.  Also, since everything is a molecule
now it is not necessary to distinguish between LIGANDS and EFFECTORS in the
count syntax so SUM_OVER_ALL_LIGANDS and SUM_OVER_ALL_EFFECTORS are also no
longer needed.  This is good news because "Simpler is Better (tm)".  As we add
counting of molecule collisions, rxns, and counting in regions we'll be adding
additional options as necessary to control the counters.  All that will be
coming soon...  So here is an example of how things look right now:

REACTION_DATA_OUTPUT {
  STEP = 1*dt
  {COUNT[A,WORLD]} => "./dat/A.dat"
  {COUNT[B,WORLD]} => "./dat/B.dat"
  {COUNT[A.B,WORLD]} => "./dat/A.B.dat"
  {COUNT[A,WORLD]+COUNT[A.B,WORLD]} => "./dat/A_A.B_tot.dat"
}

Also included in this commit are a few fixes to DX viz output and
the MDL files for the MCell3 version of the classic nmj_planar simulation.
In the process of running this sim we found a few reaction bugs that Rex
is working on and will have fixed soon.  Fortunately I don't see any molecules
leaking out (at least not within the 7ms of sim time required by the run).