Author: Rex Kerr
Fixed a bug in grid2xyz and grid2uv that was causing particles to leave
grids in peculiar places (outside the wall, potentially, for walls with
interior angles of 90+ degrees).  Also added orientation as a 5th column of
output in the CUSTOM_RK viz mode, and updated the validate_grid_util.c file
to find and fix this bug.