Author: Rex Kerr
Fixed the systematic bug--I was multiplying rather than dividing a scaling
factor which I moved from one side of an inequality to another (I divide
scaling by a random number rather than multiplying a factor by rates).

Various other stuff was changed insignificantly in the process of debugging,
and the code for preventing re-reaction immediately after release is ugly
and doesn't work for particle-particle interactions.  (That is, it's broken
and will be fixed later.) Not to mention that we still don't know what the
"right" method is to handle this.