Author: Rex Kerr
You can now use a file instead of a rate constant, as in MCell 2.  This only
works with rates, not catalysis dead time (since catalysis dead time doesn't
work itself!).  The file is interpreted as two columns of numbers; the first
is the time in seconds and the second is the rate in whatever units are
appropriate.  Any line that does not start with two numbers is assumed to be
a comment and ignored.

Note that varying rates are just stored in a linked list.  They could be put
in a scheduler, too, but that's more work.