Author: Rex Kerr
Made some changes to counter struct and updated the code to do counts of
collisions on a per-region basis.  This isn't tested (or usable), because
the code only counts when the user asks it to, and the parser has not yet
been updated to accept region-specific count commands and translate them
into the appropriate data structures.