Author: Rex Kerr
Fixed a bug in computing hash values that was causing a failure of
bimolecular reactions.  Also fixed a bug with checking for needing to count
on regions that dereferenced a null pointer.  The code was essentially
saying, in slightly disguised form, if (p==NULL) { if (p->stuff == 0) }.
Oops.

FYI, coincident reflective and transparent surfaces reflect now, unlike in
MCell2.  This will change once we figure out how this should work.