Author: Rex Kerr
Transparent surfaces now exist in two types with two syntaxes.

  TRANSPARENT = A

means that A completely ignores the surface, but not any grid molecules or
surface molecules (hereafter "effectors") on it.  It is as if it is not even
there.  If a transparent and reflective surface are coincident (reflective
is the default), you will get reflection since the transparent surface is
ignored.  (Note: the internal name for this flag is "RX_GHOST".)  This is
especially useful for, for example, boxes for counting, since you can place
them anywhere, even overlapping other surfaces, and if they are
transparent to everything the boxes will not interfere with the running of
the simulation.

However, if you say

  WINDOW = A

this means that the surface acts as a window for molecule A (an open window,
not one with glass covering it).  A will go through this without interacting
with the surface, with any other coincident surfaces, or with any effectors.
It's as if there is a hole there.

You could also have a surface transport reaction

  A' + surface_class -> A, [really_high_rate]

in which case you would go through the surface when you hit it, but you
would also interact with all the surface molecules.  This does not reliably
achieve the WINDOW property, however, since firstly, it doesn't ignore
effectors, and secondly, it has to compete with other surface reactions such
as A' + surface_class -> [high_rate] (i.e. destruction).  The MCell2
TRANSPARENT worked sort of like this, but was ill-defined if you had
multiple TRANSPARENT surfaces coincident.  If an exact copy of the MCell2
functionality is really important to have, I can create it as before
(ill-definition and all) and call this TRANSPARENT and have GHOST be the
surface property described above.

If a molecule runs into coincident surfaces--and none of them have the
WINDOW property set for that molecule--it first has a chance to react with
effectors on any of the surfaces.  (It builds a master reaction probability
table on the fly, then picks one of the available reactions.) If it still
exists and has not been transported to the other side of the surfaces, then
it has a chance to interact with any of the surfaces (not counting those
flagged TRANSPARENT to it).

Counters are updated as if all the surfaces are hit and, if appropriate,
passed through.

I don't currently do reaction counting of the special "GHOST" and "WINDOW"
reactions, as the surfaces themselves count this, and I couldn't think why
you'd want a global count of the number of times you passed through a window
or ghost/transparent surface.  But I could add it.

Finally, note that validate/mcell3_box_rx.mdl has been altered to have a
pair of internal coincident boxes for testing.  If you WINDOW, the molecules
escape from the internal box and react with the effectors (if there are any)
on the outer surface.  If you TRANSPARENT, the molecules react with
effectors on the internal box (both on the TRANSPARENT box and on the normal
reflective one).