Author: Rex Kerr
WARNING--I have changed the makefile!

I've cleaned up the makefile to do what a makefile should.  That is, when
you make a change to anything, you type "make" and everything that needs to
be recompiled or what is actually recompiled.  Previously, we were missing
the dependencies on a whole pile of header files, so whenever we changed a
header we had to "make clean" and then "make", which is kind of silly.

I have also renamed "test_main.c" as "mcell.c" and changed the output to
produce the executable mcell3.

From now on, if you ever type "make" and it doesn't make something that
needs to be made and you get a linker error or something, please fix the
makefile to include the proper dependency!

Also, watch out for the deletion of test_main.c.  Maybe I should have just
left it there, but I guess we'll find out whether CVS handles deletion as
gracefully as it should.