Author: Rex Kerr
Limited the jump-ahead time on molecules to 5000 timesteps (as a way to
perform garbage collection on the remnants of molecules that don't do
anything on their own and have been reacted with and destroyed).

Added the function is_manifold to wall_util; this will test whether or not a
region defines a manifold.  It assumes that the world has already been
sharpened (i.e. walls have been given edges), and that the region you are
testing has already been placed on the walls' region lists.