Author: Rex Kerr
Added in what I think is the rest of the computational code for counting on
and in regions.  Waypoints are to be created only if there are count
statements on regions for free molecules--we'll have to set a flag at
initialization time.  Waypoints must be set up after subvolumes are set up
and walls are distributed throughout the world--there's a function in
count_util, place_waypoints(), to do it.