Author: Rex Kerr
Fixed counting bugs inside regions (three of them, actually).  Two were just
me being stupid and not checking for hitting a subvolume boundary before
hitting a wall.  The third is rather subtle: collide_wall doesn't return the
time of collision, it returns a time slightly early!  This is to make sure
that you can never interact with something on the far side of a wall. For
counting, though, it's enough to occasionally mess up the count.  The "hit"
location is exactly right, though, so I used that instead of the time to
determine whether we'd hit too early, and now counts are fine.

Incidentally, counting of region molecules on regions also appears to be
fine.  I'm not sure what was wrong there, but whatever it was, it apparently
isn't wrong any more.

(Note for Tom: /home/kerr/spher/sphere_shells/mcell3/ has the .mdl file that
does counting on regions; the sum AChR.R over all 10 individual shells
equals the global count of AChR.R.)