Author: Rex Kerr
Found and fixed the bug that was causing crashes with counting grid
molecules on regions.  Oddly, it did not crash on my machine even though it
should have been dereferencing a NULL pointer (and it even gave the correct
answer, which should have been impossible with a NULL pointer there!).
Anyway, it crashed on Tom's laptop--and now it ought not crash.