Author: Rex Kerr
Fixed a very important bug with calculating the duration a diffusing
particle could time-jump; the code was improperly estimating the time until
it struck a wall.  If you've done a simulation that uses time-jumping during
diffusion, better check to make sure the results haven't changed!

Also, timesteps can now vary on a species-by-species basis.  Use the
  CUSTOM_TIME_STEP =
command inside a DEFINE_MOLECULE block for a particular molecule.  It must
go after the diffusion constant(s), and time is measured in seconds.

You can also automatically have all species choose a timestep so that they
travel the same mean distance each time they're updated.  In addition to the
  TIME_STEP =
command, add
  SPACE_STEP =
to set the mean diffusion step length ("l_r bar" to some of us) for all
molecules.  The time step for each species will be set accordingly (unless
you override it with a CUSTOM_TIME_STEP).

Also, you can mark species as being able to be hit, but unable to hit other
molecules.  This is supposed to be used in the case where you have a
low-abundance molecule interacting with a high-abundance one.  Since it's
faster to not have to check molecular collisions, you can let the
high-abundance one be hit but not hit other molecules.  Probabilities are
adjusted accordingly to make the rates work out.  The keyword is
  TARGET_ONLY
as the last entry in a DEFINE_MOLECULE block for a specific species.

I have not tested all of these thoroughly; I have tested to make sure that
this doesn't break some of the test-cases we already have.