Author: Rex Kerr
There was a bug in my interaction disk area estimation code that was causing
it to work improperly.  I've rewritten it somewhat.  However, the best way
to avoid errors at partition boundaries and near surfaces is to use a small
interaction radius.  This command's been in the code for some time:
  INTERACTION_RADIUS =
somewhere below the effector grid density setting.  If this is set 40x
smaller than the spacing between the partitions, error is negligible.  At 8x
smaller, the error is around 2%.  If the INTERACTION_RADIUS is not set, a
radius that makes the interaction disk area equal to the effector grid
density will be used.

Also added a message that says how many molecule-molecule collisions have
been detected at the end of each run.