Author: Rex Kerr
Added code to check whether a target diffusing molecule is on the opposite
side of a wall from the moving molecule.  Also added a lot of code to
compute the exact interaction disk area, but it's not done yet so isn't
called.

Note: this is something of a test, as it's being committed from my machine
at UCSD, not the normal one I use at the Salk.