Author: Rex Kerr
Added comments to all of the functions in the _util files that I wrote, and
hopefully updated ones that I've changed.  I did this pretty quickly, so
alert me if there are any missing comments or discrepancies between the
comment and, e.g., the arguments passed to the function.

I also cleaned up some of the memory handling in mem_util and sched_util.
The policy I have adopted is:

  A function that returns a pointer returns NULL on memory allocation
failure.  It is up to the calling function to check.
  A function that allocates memory that would otherwise return void returns
an int, which is 1 on error and 0 otherwise.
  A function that returns an non-negative integer returns -1 on memory
allocation failure.

In all cases, I've left it up to the calling function to check for being out
of memory, rather than printing something out that alerts the user but fails
to impact the execution of the program.  If we really want to print out an
error on memory allocation, we should globally switch from malloc to
something like this:
  void* warn_malloc(int n)
  {
    void* temp = malloc(n);
    if (temp==NULL) printf("Memory allocation error.\n");
    return temp;
  }
rather than duplicating this code a zillion times all over the place.

Also, mem_util will now count memory allocations and print "Nuts!" when
compiled in debug mode, but otherwise will just use straight malloc and
return NULL on error.

The reaction code has not yet been updated to be memory-error aware.  That's
next.