Author: Rex Kerr
Modified the RX_* flags in mcell_structs.h to be sensible return values for
the outcome_* set of reaction-handling functions in react_outc.c, and
updated diffuse.c to handle the memory errors that could be generated.

In diffuse.c, I have adopted the policy that the code will simply spit out
an error message and recover gracefully when it is out of memory.  This is
done with a macro called ERROR_AND_RETURN.  We may in the future wish to
simply exit the program when this occurs, or at least increment an
out-of-memory counter in struct volume so that we can print a summary for
the user at the end of the run (they will probably want to toss the entire
run if there is even a single out of memory error).

Right now the macro only uses return and not exit(1) (or more properly,
exit(EXIT_FAILURE)).