Author: Rex Kerr
I think I've fixed the last counting bug.  The bug was caused when a
waypoint was lined up with the edge of a triangle (which would happen when a
piece of geometry was exactly halfway between partition boundaries).  I've
fixed this in two ways.  First, I've shifted waypoints from the center of
subvolumes by a small fractional factor that is multiplied by powers of pi
(which should be essentially impossible to get accidentally).  Second, I've
fixed the handling of a collision with an edge to properly (hopefully)
retrace all the walls with a slightly moved target.