Author: Rex Kerr
Fixed a bug that didn't properly reflect molecules striking reflective
surfaces when counting was turned on; when counting hits (but not passes),
the code moved the molecule as if it had gone through the surface.  This
could create a hang as the molecule infinitely reflected and hopped to the
other side to reflect again.

I have only fixed the code for NON-coincident walls.  The coincident wall
code is hairy and frightening, and there is a good chance that it will be
completely rewritten without the "window" functionality (which is what makes
it so nasty).  Hopefully discussions to determine this will occur within the
next day or so; if there window is staying in I'll examine the existing code
to make sure coincident surfaces are being properly handled in this case,
and if it's not, I'll be careful to do it properly when I rewrite the code.

  --Rex