Author: Rex Kerr
Reaction counting "should be" hooked up and working, but I have not tested
it at *all*, so it is probably not working.  I'm committing now because the
rewriting led to a bug fix for counting of molecules (see below).

Worked on reaction counting a bunch.  There is a new option for counting:
  COUNT[ molecule , object/region , ALL_ENCLOSED ]
This is going to (but may not yet) count all things inside a region when you
would normally count on the region (e.g. surface molecules and reactions
that happen at a surface).

In order to make this work, I changed the counting code from taking
(molecule,number) to (molecule,number,reaction) as arguments.  The last one
is NULL if you're counting molecules; if the reaction is non-NULL then the
molecule is used only to tell where the reaction is taking place.  (This
requires knowing where grid molecules are, so I added grid_util.h to the
dependencies for count_util.c in the makefile.)

While doing this, I also fixed a bug relating to counting, where molecules
were being counted when they were exactly in a surface when produced by a
reaction on a surface.  I moved the counting code a little later to fix this
problem.  However, this introduces another bug: if you have a moving
molecule reacting directly with a surface, and that surface is full, it's
possible for products to be created by the molecule->surface reaction that
should not be, and these will not be counted.  (Before they would be
counted when erroniously produced.)

Since most people don't use reactions between molecules and surfaces (since
it's a new feature to MCell3), I'm going to ignore this bug until I
completely redo all this code again when implementing 2D diffusion.

Incidentally, I moved all the logic for setting up counters from mdlparse.y
to mdlparse_util.c.  Most of it was in common, so I condensed what would
have been five to eight separate blocks of code into a single function.
Hopefully I preserved all the important distinctions--but I may not have, so
continue to be watchful of weird counting behavior.