Author: Rex Kerr
This is a rather major update.  I'll try to remember to mention everything.
Please warn people to be watchful of their results, as I've touched a lot of
this code and although I've tested it on quite a few MDL files, I'm not
positive that I caught everything, or even remembered to do everything that
I was supposed to do.

WARNING: models that have WINDOWs in coincident boxes will now NO LONGER
WORK!  You will need to change to defining PATCHes on boxes.

Regions now have more sophisticated ways of specifying lists of elements.
ELEMENT_LIST still works.  This has an alias now,
  INCLUDE_ELEMENTS = ...
But you can now also have all elements except for the listed ones:
  EXCLUDE_ELEMENTS = ...
And you can add or remove regionsful of elements
  INCLUDE_REGION = name_of_region_on_same_object
  EXCLUDE_REGION = ...
And for boxes, you can define rectangular patches:
  INCLUDE_PATCH = [x1,y1,z1],[x2,y2,z2]
  EXCLUDE_PATCH = ...
It is an error if the patch does not lie entirely on one face of the box.
And you can mix and match these modifiers any way you like.  The PATCH
specifier can only be used when defining regions inside the box definition.

Boxes with patches now have more triangles than they used to.  I've added a
whole set of triangulation code in mdlparse_util.c to deal with this.
Basically, the box keeps track of subdivisions in the X, Y, and Z
directions, and draws a grid on the surface of the box according to these
subdivisions, and puts two triangles per panel of the grid.  These
subdivisions are stored in
  struct subdivision *sb;
inside struct polygon_object.  However, the box doesn't know what
subdivisions it will need until all its regions have been defined.
Therefore, it isn't populated with vertices and polygons until afterwards.

Because boxes are no longer simply 8 vertices, all the viz code was broken
for them.  I have changed the viz code to use the POLY_OBJ output blocks for
boxes also.  At some point, we should figure out how to sensibly switch
between the two--maybe use the old BOX_OBJ case when there are no patches
(i.e. sb->nx==2, sb->ny==2, sb->nz==2)?

Also, because boxes don't know how many walls they have until after they've
been defined, the REMOVE_ELEMENT code broke.  In addition, the
REMOVE_ELEMENT code was clunky/inflexible.  I have therefore modified it to
use the same specifiers as regions do for element list.  Thus, the command
is now
  REMOVE_ELEMENTS { /* INCLUDE/EXCLUDE patches, boxes, whatever */ }
This command actually creates a region (named objectname,REMOVED) that lists
the removed elements in it.  Its surface class is set to be a special flag
that is then checked for using a function called
  remove_gaps_from_regions
which effectively deletes this virtual region and fixes up all the other
regions and side-removal flags for that object.  This is kind of nice
because you can just completely remove a wall instead of making it
transparent if you want everything to go through it and don't want to count
in the box.

In order to support the INCLUDE and EXCLUDE functionality, I have switched
from using element_lists to using bit arrays to specify regions.  Thus, each
region contains a bit array
  struct bitarray *membership;
that says which elements of the object it belongs to are part of the region.
Using bit operations, the ELEMENT_LIST specifications are converted into a
bitmask; after initialization, region membership should be checked using
  get_bit(rp->membership,i)
instead of looping through the element_list array starting with
  rp->element_list_head
Naturally, I've gone through all the places this was done previously and
fixed it (I believe).

The struct and functions for dealing with bit arrays are in util.h and
util.c.  Briefly, you can allocate, duplicate, and delete bit arrays; you
can initialize them; you can set and check individual bits; and you can
perform bit operations on arrays.  The bit_array data structure is a little
weird (for memory efficiency's sake); the struct itself just contains two
ints, but the data for the bit array is supposed to go immediately after
that.  Thus, you need to allocate
  sizeof(struct bit_array) + (number of bits stored + 7)/8
bytes per bit array.  It's a bit odd, but with a new_bit_array function, the
weirdness is completely hidden from the user.

Polygon objects used to keep track of which sides had been removed using a
member data function
  unsigned short *side_stat;
which was not very efficient spacewise.  I have converted this over to a bit
array, and to make sure I caught all the places where it was used, I changed
it so that being set means the side is missing, rather than present.  Thus
what used to be
  if (pop->side_stat[i]==1)
is now
  if (!get_bit(pop->side_removed,1))
and I have updated the code accordingly everywhere I have noticed it.

The motivation for adding the patch code was so that we could stop
supporting the WINDOW property for molecule-surface interactions.  For a
variety of reasons that a number of us have discussed, this is an
ill-defined property and makes the diffusion code unnecessarily
horrible-looking.  Therefore, I have removed this capability from the code
entirely, along with the WINDOW keyword, leaving only TRANSPARENT.  This
allowed me to rip a couple hundred lines of code out of diffuse_3D; in
addition, I moved the proximity-detection code out of diffuse_3D into a new
function called safe_time_step which computes the maximum safe time step
(what a surprise!).  diffuse_3D is still a gigantic, hairy function, but it
is less incomprehensible to me than before, which is good.

Using Electric Fence, I found and fixed a bug in the code that generated
reactions for the WINDOW and GHOST properties (and now generates them for
the TRANSPARENT property).  I also fixed a couple other such bugs.

Of course, with all these changes there were a bunch of new dependencies to
add to the makefile.

I think that pretty much covers it.  Hopefully I've done enough testing to
ensure that the code still works after all this.