Author: Rex Kerr
Fixed a bug where a reaction rate of zero would cause a divide by zero for
the unimolecular reaction time, which would send back a value of Inf
(infinity).  This is exactly right, but since we have turned on code that
insists on crashing on divide-by-zero, I now explicitly check for it and set
the time to a very large but non-infinite number.

Also fixed a bug in diffuse_3D where counting when passing through
transparent surfaces would causes glitches in timing that could result in
molecules traveling too far and making it through solid surfaces that they
struck later.