Author: Rex Kerr
Fixed another counting bug.  It was just too difficult to push molecules to
the appropriate side of a surface inside which they were being counted, so
instead I'm removing the molecule from *all* region counters when it crosses
such a surface, and then adding it back again if it survives the timestep.
This is slow (you have to triple-raytrace each particle that encounters a
wall instead of only raytracing it once!) but it works.

It would be faster, most likely, to maintain a list of surfaces that had
been genuinely crossed, and then twiddle the counters on those as needed,
but that's more complicated.