Author: Rex Kerr
Fixed a bug having to do with the order of filling in values in various
structures in mdlparse.y that led to polygons always appearing unclosed (or,
alternativley, causing a segfault).

Cleaned up a few things stylistically while going through diffusion code
with Tom.

Changed how particles avoid walls.  Since we presumably no longer have
coincident walls, we now no longer hit and reflect off of walls "early"; the
reflection happens exactly at the wall, and which wall was hit is remembered
so you don't hit it again when you leave.  This may result in particles
being trapped between a pair of coincident reflective walls; if it does, and
we want to handle coincident reflective walls more gracefully than that
(even though there is currently no support for hitting all items on multiple
coincident reflective walls), I'll change it back.

Commented out the code that keeps track of path length on a
molecule-by-molecule basis.  I was using this for testing and can't think of
a good use for it outside of that.

Code runs slightly faster with these modifications (maybe 3%).