Author: Rex Kerr
Fixed the partition initialization code so that it starts in the middle and
goes outwards in either direction (reordered for loops so the middle is done
first).  Before it was assuming the middle had been filled in, but actually
doing the range below the middle first, which typically fails.  (This is to
make the code work as intended, and replaces Boris' fix to keep it from
crashing.)  Note: I have not tested this fix since I don't have a machine
that crashed on the code before Boris' fix.

Moved the "bisect" functions from vol_util to util, and added a "bin"
function.  They all work similarly, but return slightly different results.
In particular, bisect() and bin() return identical results if all values are
within the partition boundaries, except bin() returns a number that's one
higher.  bisect() is not guaranteed to return something sensible if the
value is outside partition boundaries, while bin() will return 0 if it's
below partition boundaries and N if it is above (where N is the number of
partitions).  bisect_near doesn't bin values but instead rounds to the
nearest point in the list (so the values act like attractors rather than
partitions).

Also updated the makefile to add the proper dependencies on util.h.

Changed my custom viz mode so that instead of outputting the position of
every molecule it writes out a histogram of molecules.  The syntax is
  MODE = CUSTOM_RK [ list of bin boundaries ] [vector to bin along]
and the output file has one line per time point that contains the iteration
number followed by tab-separated blocks that consist of the molecule
viz_state followed by the count in each bin.

Curiously, these values don't appear to add up to the number of molecules
printed by a REACTION_DATA_OUTPUT statement, even though internally they do
add up to the population for that species.  I'm not sure what is going on
there, but this code seems to work (and I need it for my simulations), so
here it is.

Note: it appears as though the code does not support having more than one
VIZ_DATA_OUTPUT block, even though the parser allows it.  This should be
fixed one way or the other at some point, but I have not fixed it now.